+++ src/primaite/game/agent/observations/nic_observations.py
@@ -1,19 +1,23 @@
 # © Crown-owned copyright 2025, Defence Science and Technology Laboratory UK
 from __future__ import annotations
 
-from typing import Dict, List, Optional
+from typing import ClassVar, Dict, List, Optional
 
 from gymnasium import spaces
 from gymnasium.core import ObsType
 
 from primaite.game.agent.observations.observations import AbstractObservation, WhereType
 from primaite.game.agent.utils import access_from_nested_dict, NOT_PRESENT_IN_STATE
+from primaite.simulator.network.nmne import NMNEConfig
 from primaite.utils.validation.ip_protocol import IPProtocol
 from primaite.utils.validation.port import Port
 
 
 class NICObservation(AbstractObservation, discriminator="network-interface"):
     """Status information about a network interface within the simulation environment."""
+
+    capture_nmne: ClassVar[bool] = NMNEConfig().capture_nmne
+    "A Boolean specifying whether malicious network events should be captured."
 
     class ConfigSchema(AbstractObservation.ConfigSchema):
         """Configuration schema for NICObservation."""
@@ -25,7 +29,13 @@ class NICObservation(AbstractObservation, discriminator="network-interface"):
         monitored_traffic: Optional[Dict[IPProtocol, List[Port]]] = None
         """A dict containing which traffic types are to be included in the observation."""
 
-    def __init__(self, where: WhereType, include_nmne: bool, monitored_traffic: Optional[Dict] = None) -> None:
+    def __init__(
+        self,
+        where: WhereType,
+        include_nmne: bool,
+        monitored_traffic: Optional[Dict] = None,
+        thresholds: Dict = {},
+    ) -> None:
         """
         Initialise a network interface observation instance.
 
@@ -45,10 +55,18 @@ class NICObservation(AbstractObservation, discriminator="network-interface"):
             self.nmne_inbound_last_step: int = 0
             self.nmne_outbound_last_step: int = 0
 
-        # TODO: allow these to be configured in yaml
-        self.high_nmne_threshold = 10
-        self.med_nmne_threshold = 5
-        self.low_nmne_threshold = 0
+        if thresholds.get("nmne") is None:
+            self.low_nmne_threshold = 0
+            self.med_nmne_threshold = 5
+            self.high_nmne_threshold = 10
+        else:
+            self._set_nmne_threshold(
+                thresholds=[
+                    thresholds.get("nmne")["low"],
+                    thresholds.get("nmne")["medium"],
+                    thresholds.get("nmne")["high"],
+                ]
+            )
 
         self.monitored_traffic = monitored_traffic
         if self.monitored_traffic:
@@ -105,6 +123,20 @@ class NICObservation(AbstractObservation, discriminator="network-interface"):
         bandwidth_utilisation = traffic_value / nic_max_bandwidth
         return int(bandwidth_utilisation * 9) + 1
 
+    def _set_nmne_threshold(self, thresholds: List[int]):
+        """
+        Method that validates and then sets the NMNE threshold.
+
+        :param: thresholds: The NMNE threshold to validate and set.
+        """
+        if self._validate_thresholds(
+            thresholds=thresholds,
+            threshold_identifier="nmne",
+        ):
+            self.low_nmne_threshold = thresholds[0]
+            self.med_nmne_threshold = thresholds[1]
+            self.high_nmne_threshold = thresholds[2]
+
     def observe(self, state: Dict) -> ObsType:
         """
         Generate observation based on the current state of the simulation.
@@ -116,7 +148,7 @@ class NICObservation(AbstractObservation, discriminator="network-interface"):
         """
         nic_state = access_from_nested_dict(state, self.where)
 
-        if nic_state is NOT_PRESENT_IN_STATE:
+        if nic_state is NOT_PRESENT_IN_STATE or self.where is None:
             return self.default_observation
 
         obs = {"nic_status": 1 if nic_state["enabled"] else 2}
@@ -164,7 +196,7 @@ class NICObservation(AbstractObservation, discriminator="network-interface"):
                         for port in self.monitored_traffic[protocol]:
                             obs["TRAFFIC"][protocol][port] = {"inbound": 0, "outbound": 0}
 
-        if self.include_nmne:
+        if self.capture_nmne and self.include_nmne:
             obs.update({"NMNE": {}})
             direction_dict = nic_state["nmne"].get("direction", {})
             inbound_keywords = direction_dict.get("inbound", {}).get("keywords", {})
@@ -224,6 +256,7 @@ class NICObservation(AbstractObservation, discriminator="network-interface"):
             where=parent_where + ["NICs", config.nic_num],
             include_nmne=config.include_nmne,
             monitored_traffic=config.monitored_traffic,
+            thresholds=config.thresholds,
         )