+++ src/primaite/game/agent/observations/node_observations.py
@@ -48,7 +48,13 @@ class NodesObservation(AbstractObservation, discriminator="nodes"):
         include_num_access: Optional[bool] = None
         """Flag to include the number of accesses."""
         file_system_requires_scan: bool = True
-        """If True, the folder must be scanned to update the health state. Tf False, the true state is always shown."""
+        """If True, the folder must be scanned to update the health state. If False, the true state is always shown."""
+        services_requires_scan: bool = True
+        """If True, the services must be scanned to update the health state.
+        If False, the true state is always shown."""
+        applications_requires_scan: bool = True
+        """If True, the applications must be scanned to update the health state.
+        If False, the true state is always shown."""
         include_users: Optional[bool] = True
         """If True, report user session information."""
         num_ports: Optional[int] = None
@@ -196,8 +202,14 @@ class NodesObservation(AbstractObservation, discriminator="nodes"):
                 host_config.include_num_access = config.include_num_access
             if host_config.file_system_requires_scan is None:
                 host_config.file_system_requires_scan = config.file_system_requires_scan
+            if host_config.services_requires_scan is None:
+                host_config.services_requires_scan = config.services_requires_scan
+            if host_config.applications_requires_scan is None:
+                host_config.applications_requires_scan = config.applications_requires_scan
             if host_config.include_users is None:
                 host_config.include_users = config.include_users
+            if not host_config.thresholds:
+                host_config.thresholds = config.thresholds
 
         for router_config in config.routers:
             if router_config.num_ports is None:
@@ -214,6 +226,8 @@ class NodesObservation(AbstractObservation, discriminator="nodes"):
                 router_config.num_rules = config.num_rules
             if router_config.include_users is None:
                 router_config.include_users = config.include_users
+            if not router_config.thresholds:
+                router_config.thresholds = config.thresholds
 
         for firewall_config in config.firewalls:
             if firewall_config.ip_list is None:
@@ -228,6 +242,8 @@ class NodesObservation(AbstractObservation, discriminator="nodes"):
                 firewall_config.num_rules = config.num_rules
             if firewall_config.include_users is None:
                 firewall_config.include_users = config.include_users
+            if not firewall_config.thresholds:
+                firewall_config.thresholds = config.thresholds
 
         hosts = [HostObservation.from_config(config=c, parent_where=where) for c in config.hosts]
         routers = [RouterObservation.from_config(config=c, parent_where=where) for c in config.routers]