+++ tests/integration_tests/game_layer/observations/test_software_observations.py
@@ -29,7 +29,9 @@ def test_service_observation(simulation):
     ntp_server = pc.software_manager.software.get("ntp-server")
     assert ntp_server
 
-    service_obs = ServiceObservation(where=["network", "nodes", pc.config.hostname, "services", "ntp-server"])
+    service_obs = ServiceObservation(
+        where=["network", "nodes", pc.config.hostname, "services", "ntp-server"], services_requires_scan=True
+    )
 
     assert service_obs.space["operating_status"] == spaces.Discrete(7)
     assert service_obs.space["health_status"] == spaces.Discrete(5)
@@ -54,7 +56,9 @@ def test_application_observation(simulation):
     web_browser: WebBrowser = pc.software_manager.software.get("web-browser")
     assert web_browser
 
-    app_obs = ApplicationObservation(where=["network", "nodes", pc.config.hostname, "applications", "web-browser"])
+    app_obs = ApplicationObservation(
+        where=["network", "nodes", pc.config.hostname, "applications", "web-browser"], applications_requires_scan=True
+    )
 
     web_browser.close()
     observation_state = app_obs.observe(simulation.describe_state())
@@ -69,3 +73,33 @@ def test_application_observation(simulation):
     assert observation_state.get("health_status") == 1
     assert observation_state.get("operating_status") == 1  # running
     assert observation_state.get("num_executions") == 1
+
+
+def test_application_executions_categories(simulation):
+    pc: Computer = simulation.network.get_node_by_hostname("client_1")
+
+    app_obs = ApplicationObservation(
+        where=["network", "nodes", pc.config.hostname, "applications", "WebBrowser"],
+        applications_requires_scan=False,
+        thresholds={"app_executions": {"low": 3, "medium": 6, "high": 9}},
+    )
+
+    assert app_obs.high_app_execution_threshold == 9
+    assert app_obs.med_app_execution_threshold == 6
+    assert app_obs.low_app_execution_threshold == 3
+
+    with pytest.raises(Exception):
+        # should throw an error
+        ApplicationObservation(
+            where=["network", "nodes", pc.config.hostname, "applications", "WebBrowser"],
+            applications_requires_scan=False,
+            thresholds={"app_executions": {"low": 9, "medium": 6, "high": 9}},
+        )
+
+    with pytest.raises(Exception):
+        # should throw an error
+        ApplicationObservation(
+            where=["network", "nodes", pc.config.hostname, "applications", "WebBrowser"],
+            applications_requires_scan=False,
+            thresholds={"app_executions": {"low": 3, "medium": 9, "high": 9}},
+        )